misc: split dialects.utils into multiple files
