+++ xdsl/dialects/linalg.py
@@ -25,7 +25,7 @@ from xdsl.dialects.builtin import (
     TensorType,
     i64,
 )
-from xdsl.dialects.utils import (
+from xdsl.dialects.utils.format import (
     AbstractYieldOperation,
 )
 from xdsl.ir import (
