+++ xdsl/dialects/riscv.py
@@ -24,7 +24,7 @@ from xdsl.dialects.builtin import (
     UnitAttr,
     i32,
 )
-from xdsl.dialects.utils import FastMathAttrBase, FastMathFlag
+from xdsl.dialects.utils.fast_math import FastMathAttrBase, FastMathFlag
 from xdsl.ir import (
     Attribute,
     Block,
